Added features to Course class
Added the finalCalculate class to Course, so class instances will now be
able to call their own final calculate method.  Also added a constructor
to take a boolean flag to determine weather the class is out of points
or percent.  Calling finalCalculate will automatically calculate final
grade needed for points or percent, depending on what the class grade
distribution is set to.