Assignment 9
The AttributeOfClass.java demonstrates the Single responsibility
principle, where a class should only have responsibility over a single
part of the functionality provided by the software.  I pushed this code
in the pervious push as well, but I've made some changes in this
version.  This helps improve the quality of the code by making things
more straight forward for maintenance purposes, as well as making the
debug process smoother.